# Improve performance of feed_manager_spec