In order to avoid creating many records, I changed the value of the constant to stub in the test.

before:
```
bundle exec rspec spec/lib/feed_manager_spec.rb

Randomized with seed 22807
 32/32 |============================================================== 100 ==============================================================>| Time: 00:00:46

Finished in 46.71 seconds (files took 4.32 seconds to load)
32 examples, 0 failures
```

after:
```
bundle exec rspec spec/lib/feed_manager_spec.rb

Randomized with seed 27571
 32/32 |============================================================== 100 ==============================================================>| Time: 00:00:08

Finished in 8.3 seconds (files took 4.31 seconds to load)
32 examples, 0 failures
```